requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
opencv-python==4.8.1.78
Pillow==11.3.0
insightface==0.7.3
onnxruntime==1.16.3
qdrant-client==1.7.0
numpy==1.24.3
pydantic==2.5.0
python-multipart==0.0.6
httpx==0.25.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4 